requirements.txt
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.6.4
schedule==1.2.0
